requirements.txt
## Python Version = 3.12.4
Flask
pymongo
bcrypt
numpy
plotly
dash
flask_wtf
wtforms
wtforms.validators
dash_bootstrap_components
pandas
Werkzeug
email_validator
certifi
reportLab
pdfrw
pyPDF2
python-dotenv
psycopg2
streamlit
seaborn
matplotlib